requirements.txt
requests
icecream
loguru
bs4
lxml
python-dotenv
urllib3
aiohttp
aiofiles
fake-useragent
python-telegram-bot==13.15
